surl/requirements/base.txt
alembic
asyncpg
backoff
bcrypt
fastapi
fastapi-utils
httpx
itsdangerous
passlib
psycopg2-binary
pydantic
pytest
pytest-asyncio
pytest-cov
python-jose
python-multipart
rich
sqlalchemy
sqlalchemy-utils
sqlmodel
tenacity
typer[all]
uvicorn
